#80 feat: 이미지, 동영상 res -> aws s3 bucket 주소로 변경
