+++ src/main/java/com/meongmory/meongmory/domain/diary/dto/response/FileRes.java
@@ -1,8 +1,10 @@
 package com.meongmory.meongmory.domain.diary.dto.response;
 
 import com.meongmory.meongmory.domain.diary.entity.DiaryFile;
+import com.meongmory.meongmory.global.util.AwsS3ImageUrlUtil;
 import lombok.Builder;
 import lombok.Data;
+import lombok.RequiredArgsConstructor;
 
 import java.util.List;
 
@@ -20,8 +22,8 @@ public class FileRes {
 
   public static FileRes toDto(DiaryFile diaryFile) {
     return FileRes.builder()
-        .fileKey(diaryFile.getFileKey())
-        .fileType(diaryFile.getType().name())
-        .build();
+            .fileKey(AwsS3ImageUrlUtil.toUrl(diaryFile.getFileKey()))
+            .fileType(diaryFile.getType().name())
+            .build();
   }
 }
