+++ src/main/java/com/meongmory/meongmory/domain/diary/entity/Diary.java
@@ -23,6 +23,9 @@ public class Diary extends BaseEntity {
   @OneToMany(mappedBy = "diary", cascade = CascadeType.ALL)
   private List<DiaryFile> files = new ArrayList<>();
 
+  @OneToMany(mappedBy = "pet", cascade = CascadeType.ALL)
+  private List<DiaryPet> diaryPets = new ArrayList<>();
+
   @ManyToOne(fetch = FetchType.LAZY, optional = false)
   @JoinColumn(name="familyId")
   private Family family;
